Broke out all functions into their own files and tidied up. Moved credit card types into a json file for easy editting
