+++ index.js
@@ -1,453 +1,5 @@
 'use strict';
 
-var testOrder;
-var types = {};
-var customCards = {};
-var VISA = 'visa';
-var MASTERCARD = 'mastercard';
-var AMERICAN_EXPRESS = 'american-express';
-var DINERS_CLUB = 'diners-club';
-var DISCOVER = 'discover';
-var ELO = 'elo';
-var JCB = 'jcb';
-var UNIONPAY = 'unionpay';
-var MAESTRO = 'maestro';
-var MIR = 'mir';
-var CVV = 'CVV';
-var CID = 'CID';
-var CVC = 'CVC';
-var CVN = 'CVN';
-var CVP2 = 'CVP2';
-var CVE = 'CVE';
-var ORIGINAL_TEST_ORDER = [
-  VISA,
-  MASTERCARD,
-  AMERICAN_EXPRESS,
-  DINERS_CLUB,
-  DISCOVER,
-  JCB,
-  UNIONPAY,
-  MAESTRO,
-  ELO,
-  MIR
-];
-
-function clone(originalObject) {
-  var dupe;
-
-  if (!originalObject) { return null; }
-
-  dupe = JSON.parse(JSON.stringify(originalObject));
-
-  return dupe;
-}
-
-testOrder = clone(ORIGINAL_TEST_ORDER);
-
-types[VISA] = {
-  niceType: 'Visa',
-  type: VISA,
-  patterns: [
-    4
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16, 18, 19],
-  code: {
-    name: CVV,
-    size: 3
-  }
-};
-
-types[MASTERCARD] = {
-  niceType: 'Mastercard',
-  type: MASTERCARD,
-  patterns: [
-    [51, 55],
-    [2221, 2229],
-    [223, 229],
-    [23, 26],
-    [270, 271],
-    2720
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16],
-  code: {
-    name: CVC,
-    size: 3
-  }
-};
-
-types[AMERICAN_EXPRESS] = {
-  niceType: 'American Express',
-  type: AMERICAN_EXPRESS,
-  patterns: [
-    34,
-    37
-  ],
-  gaps: [4, 10],
-  lengths: [15],
-  code: {
-    name: CID,
-    size: 4
-  }
-};
-
-types[DINERS_CLUB] = {
-  niceType: 'Diners Club',
-  type: DINERS_CLUB,
-  patterns: [
-    [300, 305],
-    36,
-    38,
-    39
-  ],
-  gaps: [4, 10],
-  lengths: [14, 16, 19],
-  code: {
-    name: CVV,
-    size: 3
-  }
-};
-
-types[DISCOVER] = {
-  niceType: 'Discover',
-  type: DISCOVER,
-  patterns: [
-    6011,
-    [644, 649],
-    65
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16, 19],
-  code: {
-    name: CID,
-    size: 3
-  }
-};
-
-types[JCB] = {
-  niceType: 'JCB',
-  type: JCB,
-  patterns: [
-    2131,
-    1800,
-    [3528, 3589]
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16, 17, 18, 19],
-  code: {
-    name: CVV,
-    size: 3
-  }
-};
-
-types[UNIONPAY] = {
-  niceType: 'UnionPay',
-  type: UNIONPAY,
-  patterns: [
-    620,
-    [624, 626],
-    [62100, 62182],
-    [62184, 62187],
-    [62185, 62197],
-    [62200, 62205],
-    [622010, 622999],
-    622018,
-    [622019, 622999],
-    [62207, 62209],
-    [622126, 622925],
-    [623, 626],
-    6270,
-    6272,
-    6276,
-    [627700, 627779],
-    [627781, 627799],
-    [6282, 6289],
-    6291,
-    6292
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16, 17, 18, 19],
-  code: {
-    name: CVN,
-    size: 3
-  }
-};
-
-types[MAESTRO] = {
-  niceType: 'Maestro',
-  type: MAESTRO,
-  patterns: [
-    493698,
-    [500000, 506698],
-    [506779, 508999],
-    [56, 59],
-    63,
-    67,
-    6
-  ],
-  gaps: [4, 8, 12],
-  lengths: [12, 13, 14, 15, 16, 17, 18, 19],
-  code: {
-    name: CVC,
-    size: 3
-  }
-};
-
-types[ELO] = {
-  niceType: 'Elo',
-  type: ELO,
-  patterns: [
-    401178,
-    401179,
-    438935,
-    457631,
-    457632,
-    431274,
-    451416,
-    457393,
-    504175,
-    [506699, 506778],
-    [509000, 509999],
-    627780,
-    636297,
-    636368,
-    [650031, 650033],
-    [650035, 650051],
-    [650405, 650439],
-    [650485, 650538],
-    [650541, 650598],
-    [650700, 650718],
-    [650720, 650727],
-    [650901, 650978],
-    [651652, 651679],
-    [655000, 655019],
-    [655021, 655058]
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16],
-  code: {
-    name: CVE,
-    size: 3
-  }
-};
-
-types[MIR] = {
-  niceType: 'Mir',
-  type: MIR,
-  patterns: [
-    [2200, 2204]
-  ],
-  gaps: [4, 8, 12],
-  lengths: [16, 17, 18, 19],
-  code: {
-    name: CVP2,
-    size: 3
-  }
-};
-
-function findType(type) {
-  return customCards[type] || types[type];
-}
-
-function isValidInputType(cardNumber) {
-  return typeof cardNumber === 'string' || cardNumber instanceof String;
-}
-
-function hasEnoughResultsToDetermineBestMatch(results) {
-  var numberOfResultsWithMaxStrengthProperty = results.filter(function (result) {
-    return result.matchStrength;
-  }).length;
-
-  // if all possible results have a maxStrength property
-  // that means the card number is sufficiently long
-  // enough to determine conclusively what the type is
-  return numberOfResultsWithMaxStrengthProperty > 0 &&
-    numberOfResultsWithMaxStrengthProperty === results.length;
-}
-
-function findBestMatch(results) {
-  if (!hasEnoughResultsToDetermineBestMatch(results)) {
-    return;
-  }
-
-  return results.reduce(function (bestMatch, result) { // eslint-disable-line consistent-return
-    if (!bestMatch) {
-      return result;
-    }
-
-    // if the current best match pattern is less specific
-    // than this result, set the result as the new best match
-    if (bestMatch.matchStrength < result.matchStrength) {
-      return result;
-    }
-
-    return bestMatch;
-  });
-}
-
-function getAllCardTypes() {
-  return testOrder.map(function (type) {
-    return clone(findType(type));
-  });
-}
-
-function creditCardType(cardNumber) {
-  var bestMatch;
-  var results = [];
-
-  if (!isValidInputType(cardNumber)) {
-    return [];
-  }
-
-  if (cardNumber.length === 0) {
-    return getAllCardTypes();
-  }
-
-  testOrder.forEach(function (type) {
-    var cardConfiguration = findType(type);
-
-    loopOverCardPatterns(cardNumber, cardConfiguration, results);
-  });
-
-  bestMatch = findBestMatch(results);
-
-  if (bestMatch) {
-    return [bestMatch];
-  }
-
-  return results;
-}
-
-function loopOverCardPatterns(cardNumber, cardConfiguration, results) {
-  var i, pattern, patternLength, clonedCardConfiguration;
-
-  for (i = 0; i < cardConfiguration.patterns.length; i++) {
-    pattern = cardConfiguration.patterns[i];
-
-    if (!matches(cardNumber, pattern)) {
-      continue;
-    }
-
-    clonedCardConfiguration = clone(cardConfiguration);
-
-    if (Array.isArray(pattern)) {
-      patternLength = String(pattern[0]).length;
-    } else {
-      patternLength = String(pattern).length;
-    }
-
-    if (cardNumber.length >= patternLength) {
-      clonedCardConfiguration.matchStrength = patternLength;
-    }
-
-    results.push(clonedCardConfiguration);
-    break;
-  }
-}
-
-function matches(cardNumber, pattern) {
-  if (Array.isArray(pattern)) {
-    return matchesRange(cardNumber, pattern[0], pattern[1]);
-  }
-
-  return matchesPattern(cardNumber, pattern);
-}
-
-function matchesPattern(cardNumber, pattern) {
-  pattern = String(pattern);
-
-  return pattern.substring(0, cardNumber.length) === cardNumber.substring(0, pattern.length);
-}
-
-// Adapted from https://github.com/polvo-labs/card-type/blob/aaab11f80fa1939bccc8f24905a06ae3cd864356/src/cardType.js#L37-L42
-function matchesRange(cardNumber, min, max) {
-  var maxLengthToCheck = String(min).length;
-  var substr = cardNumber.substr(0, maxLengthToCheck);
-  var integerRepresentationOfCardNumber = parseInt(substr, 10);
-
-  min = parseInt(String(min).substr(0, substr.length), 10);
-  max = parseInt(String(max).substr(0, substr.length), 10);
-
-  return integerRepresentationOfCardNumber >= min && integerRepresentationOfCardNumber <= max;
-}
-
-creditCardType.getTypeInfo = function (type) {
-  return clone(findType(type));
-};
-
-function getCardPosition(name, ignoreErrorForNotExisting) {
-  var position = testOrder.indexOf(name);
-
-  if (!ignoreErrorForNotExisting && position === -1) {
-    throw new Error('"' + name + '" is not a supported card type.');
-  }
-
-  return position;
-}
-
-creditCardType.removeCard = function (name) {
-  var position = getCardPosition(name);
-
-  testOrder.splice(position, 1);
-};
-
-creditCardType.addCard = function (config) {
-  var existingCardPosition = getCardPosition(config.type, true);
-
-  customCards[config.type] = config;
-
-  if (existingCardPosition === -1) {
-    testOrder.push(config.type);
-  }
-};
-
-creditCardType.updateCard = function (cardType, updates) {
-  var clonedCard;
-  var originalObject = customCards[cardType] || types[cardType];
-
-  if (!originalObject) {
-    throw new Error('"' + cardType + '" is not a recognized type. Use `addCard` instead.');
-  }
-
-  if (updates.type && originalObject.type !== updates.type) {
-    throw new Error('Cannot overwrite type parameter.');
-  }
-
-  clonedCard = clone(originalObject, true);
-
-  Object.keys(clonedCard).forEach(function (key) {
-    if (updates[key]) {
-      clonedCard[key] = updates[key];
-    }
-  });
-
-  customCards[clonedCard.type] = clonedCard;
-};
-
-creditCardType.changeOrder = function (name, position) {
-  var currentPosition = getCardPosition(name);
-
-  testOrder.splice(currentPosition, 1);
-  testOrder.splice(position, 0, name);
-};
-
-creditCardType.resetModifications = function () {
-  testOrder = clone(ORIGINAL_TEST_ORDER);
-  customCards = {};
-};
-
-creditCardType.types = {
-  VISA: VISA,
-  MASTERCARD: MASTERCARD,
-  AMERICAN_EXPRESS: AMERICAN_EXPRESS,
-  DINERS_CLUB: DINERS_CLUB,
-  DISCOVER: DISCOVER,
-  JCB: JCB,
-  UNIONPAY: UNIONPAY,
-  MAESTRO: MAESTRO,
-  ELO: ELO,
-  MIR: MIR
-};
+var creditCardType = require('./validator/creditCardType');
 
 module.exports = creditCardType;
