+++ validator/creditCardType.js
@@ -1,0 +1,135 @@
+'use strict';
+
+var types = require('./cardTypes.json');
+var clone = require('./helpers/clone');
+var findBestMatch = require('./helpers/findBestMatch');
+var findType = require('./helpers/findType');
+var getAllCardTypes = require('./helpers/getAllCardTypes');
+var getCardPosition = require('./helpers/getCardPosition');
+var isValidInputType = require('./helpers/isValidInputType');
+var loopOverCardPatterns = require('./helpers/loopOverCardPatterns');
+
+var testOrder;
+var customCards = {};
+var VISA = 'visa';
+var MASTERCARD = 'mastercard';
+var AMERICAN_EXPRESS = 'american-express';
+var DINERS_CLUB = 'diners-club';
+var DISCOVER = 'discover';
+var ELO = 'elo';
+var JCB = 'jcb';
+var UNIONPAY = 'unionpay';
+var MAESTRO = 'maestro';
+var MIR = 'mir';
+
+var ORIGINAL_TEST_ORDER = [
+  VISA,
+  MASTERCARD,
+  AMERICAN_EXPRESS,
+  DINERS_CLUB,
+  DISCOVER,
+  JCB,
+  UNIONPAY,
+  MAESTRO,
+  ELO,
+  MIR
+];
+
+testOrder = clone(ORIGINAL_TEST_ORDER);
+
+function creditCardType(cardNumber) {
+  var bestMatch;
+  var results = [];
+
+  if (!isValidInputType(cardNumber)) {
+    return [];
+  }
+
+  if (cardNumber.length === 0) {
+    return getAllCardTypes(testOrder, customCards);
+  }
+
+  testOrder.forEach(function (type) {
+    var cardConfiguration = findType(type, customCards);
+
+    loopOverCardPatterns(cardNumber, cardConfiguration, results);
+  });
+
+  bestMatch = findBestMatch(results);
+
+  if (bestMatch) {
+    return [bestMatch];
+  }
+
+  return results;
+}
+
+creditCardType.getTypeInfo = function (type) {
+  return clone(findType(type, customCards));
+};
+
+creditCardType.removeCard = function (name) {
+  var position = getCardPosition(name, false, testOrder);
+
+  testOrder.splice(position, 1);
+};
+
+creditCardType.addCard = function (config) {
+  var existingCardPosition = getCardPosition(config.type, true, testOrder);
+
+  customCards[config.type] = config;
+
+  if (existingCardPosition === -1) {
+    testOrder.push(config.type);
+  }
+};
+
+creditCardType.updateCard = function (cardType, updates) {
+  var clonedCard;
+  var originalObject = customCards[cardType] || types[cardType];
+
+  if (!originalObject) {
+    throw new Error('"' + cardType + '" is not a recognized type. Use `addCard` instead.');
+  }
+
+  if (updates.type && originalObject.type !== updates.type) {
+    throw new Error('Cannot overwrite type parameter.');
+  }
+
+  clonedCard = clone(originalObject, true);
+
+  Object.keys(clonedCard).forEach(function (key) {
+    if (updates[key]) {
+      clonedCard[key] = updates[key];
+    }
+  });
+
+  customCards[clonedCard.type] = clonedCard;
+};
+
+creditCardType.changeOrder = function (name, position) {
+  var currentPosition = getCardPosition(name, false, testOrder);
+
+  testOrder.splice(currentPosition, 1);
+  testOrder.splice(position, 0, name);
+};
+
+creditCardType.resetModifications = function () {
+  testOrder = clone(ORIGINAL_TEST_ORDER);
+  customCards = {};
+};
+
+creditCardType.types = {
+  VISA: VISA,
+  MASTERCARD: MASTERCARD,
+  AMERICAN_EXPRESS: AMERICAN_EXPRESS,
+  DINERS_CLUB: DINERS_CLUB,
+  DISCOVER: DISCOVER,
+  JCB: JCB,
+  UNIONPAY: UNIONPAY,
+  MAESTRO: MAESTRO,
+  ELO: ELO,
+  MIR: MIR
+};
+
+module.exports = creditCardType;
