+++ validator/helpers/findBestMatch.js
@@ -1,0 +1,25 @@
+'use strict';
+
+var hasEnoughResultsToDetermineBestMatch = require('./hasEnoughResultsToDetermineBestMatch');
+
+function findBestMatch(results) {
+  if (!hasEnoughResultsToDetermineBestMatch(results)) {
+    return;
+  }
+
+  return results.reduce(function (bestMatch, result) { // eslint-disable-line consistent-return
+    if (!bestMatch) {
+      return result;
+    }
+
+    // if the current best match pattern is less specific
+    // than this result, set the result as the new best match
+    if (bestMatch.matchStrength < result.matchStrength) {
+      return result;
+    }
+
+    return bestMatch;
+  });
+}
+
+module.exports = findBestMatch;
