+++ validator/helpers/isValidInputType.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function isValidInputType(cardNumber) {
+  return typeof cardNumber === 'string' || cardNumber instanceof String;
+}
+
+module.exports = isValidInputType;
