+++ validator/helpers/matches.js
@@ -1,0 +1,14 @@
+'use strict';
+
+var matchesRange = require('./matchesRange');
+var matchesPattern = require('./matchesPattern');
+
+function matches(cardNumber, pattern) {
+  if (Array.isArray(pattern)) {
+    return matchesRange(cardNumber, pattern[0], pattern[1]);
+  }
+
+  return matchesPattern(cardNumber, pattern);
+}
+
+module.exports = matches;
