+++ validator/helpers/matchesPattern.js
@@ -1,0 +1,9 @@
+'use strict';
+
+function matchesPattern(cardNumber, pattern) {
+  pattern = String(pattern);
+
+  return pattern.substring(0, cardNumber.length) === cardNumber.substring(0, pattern.length);
+}
+
+module.exports = matchesPattern;
